add bull queue
